+++ packages/vendure-plugin-invoices/src/strategies/google-storage-invoice-strategy.ts
@@ -3,7 +3,7 @@ import { Response } from 'express';
 import { createReadStream, ReadStream } from 'fs';
 import path from 'path';
 import { InvoiceEntity } from '../entities/invoice.entity';
-import { createTempFile, zipFiles, ZippableFile } from '../util/file.util';
+import { createTempFile } from '../util/file.util';
 import { RemoteStorageStrategy } from './storage-strategy';
 
 interface GoogleInvoiceConfig {
@@ -61,28 +61,4 @@ export class GoogleStorageInvoiceStrategy implements RemoteStorageStrategy {
     return name;
   }
 
-  async streamMultiple(
-    invoices: InvoiceEntity[],
-    res: Response
-  ): Promise<ReadStream> {
-    res.set({
-      'Content-Type': 'application/zip',
-      'Content-Disposition': `inline; filename="invoices-${invoices.length}.zip"`,
-    });
-    const files: ZippableFile[] = await Promise.all(
-      invoices.map(async (invoice) => {
-        const tmpFile = await createTempFile('.pdf');
-        await this.storage
-          .bucket(this.bucketName)
-          .file(invoice.storageReference)
-          .download({ destination: tmpFile });
-        return {
-          path: tmpFile,
-          name: invoice.invoiceNumber + '.pdf',
-        };
-      })
-    );
-    const zipFile = await zipFiles(files);
-    return createReadStream(zipFile);
-  }
 }