+++ packages/vendure-plugin-invoices/src/util/file.util.ts
@@ -1,6 +1,5 @@
+import { promises as fs } from 'fs';
 import * as tmp from 'tmp';
-import { promises as fs, createWriteStream, createReadStream } from 'fs';
-import AdmZip = require('adm-zip');
 
 export async function createTempFile(postfix: string): Promise<string> {
   return new Promise((resolve, reject) => {
@@ -14,20 +13,6 @@ export async function createTempFile(postfix: string): Promise<string> {
   });
 }
 
-export interface ZippableFile {
-  name: string;
-  path: string;
-}
-
-export async function zipFiles(files: ZippableFile[]): Promise<string> {
-  const zip = new AdmZip();
-  for (const file of files) {
-    zip.addLocalFile(file.path, undefined, file.name);
-  }
-  const tmpFilePath = await createTempFile('.zip');
-  await zip.writeZip(tmpFilePath);
-  return tmpFilePath;
-}
 
 export async function exists(path: string): Promise<boolean> {
   let exists = false;