+++ packages/vendure-plugin-invoices/CHANGELOG.md
@@ -1,4 +1,11 @@
-# 1.2.0 (2023-10-24)
+# 2.0.0 (2024-01-07)
+
+- BREAKING: Script needs to be run
+- BREAKING: Invoice download URL's from before v2 will always download the first invoice of an order
+- BREAKING: Downloading multiple invoices has been removed. Invoices are downloaded on a per order basis
+- BREAKING: Downloading multiple invoices has been removed. Invoices are downloaded on a per order basis
+  
+   1.2.0 (2023-10-24)
 
 - Updated vendure to 2.1.1